requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Bcrypt==1.0.1
gunicorn==21.2.0
reportlab==4.0.4
matplotlib==3.7.2
numpy==1.24.4
openai==1.3.8
python-dotenv==1.0.0
httpx==0.25.2